Add uploading

+++ src/main/java/com/github/vacxe/apkpublisher/models/ApkInfo.kt
@@ -1,5 +1,7 @@
 package com.github.vacxe.apkpublisher.models
 
-data class ApkInfo(val name: String,
-                   val versionId: Int,
-                   val versionName: String)
+data class ApkInfo(
+    val name: String,
+    val versionId: Long,
+    val versionName: String
+)
